Refactor Sync Metadata for Alluxio engine
